% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchbook.R
\docType{data}
\name{TopSketchbook}
\alias{TopSketchbook}
\title{Top solutions to sketchbook puzzles with a fixed number of moves.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 71 rows and 7 columns.}
\usage{
TopSketchbook
}
\description{
\describe{
\item{uid}{factor Unique identifier for users.}
\item{gid}{factor Unique identifier for groups. NA for soloists.}
\item{timestamp}{int Time at which this solution was submitted.}
\item{energy}{double Energy score of winning protein configuration.}
\item{puzzle_id}{factor Unique identifier for puzzles.}
\item{prefix}{char Type of rank comparison.}
\item{rank}{int Ranked energy score by prefix.}
}
}
\keyword{datasets}
